requirements.txt
openai>=1.50.0
networkx>=3.2
matplotlib>=3.8
neo4j>=5.20
numpy>=1.24
python-dotenv>=1.0.0
pandas>=2.0
tiktoken>=0.7.0
tqdm>=4.66
